Deleting Satoshi clause
Issuing virtual currency should not be a licensed activity. Centralised virtual/digital currencies (such as fiat currency) can be licensed, but issuing a new decentralised digital currency should not be a license-able activity, as the issuers have no control over how the currency is used and cannot keep a register of users or active/ongoing AML policy on all users of the decentralised virtual currency.