Changing definition of Virtual Currency Business
Receiving virtual currency for transmitting it on to someone else is a dangerous expansion of scope. A lawyer, hedge fund, bank could easily fall into this category - they should be covered by existing banking/legal/fund regulations, not additional Bitlicense regs. Same goes for holding/storing/securing bitcoin on behalf of others. Should a law firm that holds escrow or stores bitcoin for someone need a bitlicense, when they already have a license to practice law? Should an escrow agency or arbitration agency need a bitlicense? There are a number of other similar examples of where this will apply.