Added link at top of master (fixed) proposal
This link will show the diff between the "master" branch and "original" branch of this repository. Making it very easy to see the changes.

The README.md on the "original" branch should contain the original document, and the README.md on the "master" branch should have all the proposed fixes from everyone.

This way the red/green highlighted differences will be easily accessible to anyone, including Lawsky.